pyproject.toml
[project]
name = "hagrid-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.21",
    "fastapi[standard]>=0.128.0",
    "fyers-apiv3>=3.1.7",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "setuptools>=80.9.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "jupyterlab>=4.5.1",
    "plotly>=6.5.1",
    "markdownify>=1.2.2",
    "yfinance>=1.0",
    "lxml>=6.0.2",
    "google-genai>=1.57.0",
    "scipy>=1.17.0",
    "numpy>=2.4.0",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "openinference-instrumentation-agno>=0.1.25",
]
